fixed segfault in MMU::load_cartridge

+++ core/mmu.h
@@ -10,7 +10,7 @@ enum PPUState { mode0 = 0, mode1, mode2, mode3 };
 
 struct Cartridge {
     std::array<u8, 0x50> header = {0};
-    u16 rom_size; 
+    u32 rom_size; 
     bool ram;
     u8 ram_size;
     bool battery;
